seperate Fafile in seperate module + netcdf4 engine
